requirements.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
billiard==4.2.1
build==1.2.2.post1
CacheControl==0.14.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
cryptography==44.0.0
distlib==0.3.9
dulwich==0.21.7
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
greenlet==3.1.1
ghostscript==0.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.2
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyMySQL==1.1.1
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-asyncio==0.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
RapidFuzz==3.11.0
redis==5.2.1
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
sentry-sdk==2.19.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
watchdog==6.0.0
xattr==1.1.0
zipp==3.21.0

